🔧 Some refactoring
