+++ src/client/components/Settings/QbittorrentSettings/QbittorrentConnectionForm.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement, useCallback, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { getQBitTorrentClientInfo } from "../../actions/settings.actions";
+import { getQBitTorrentClientInfo } from "../../../actions/settings.actions";
 
 export const QbittorrentConnectionForm = (): ReactElement => {
   const dispatch = useDispatch();
